# Faster escape filter (for v2)