--- a/plugins/escape.ts
+++ b/plugins/escape.ts
@@ -1,21 +1,47 @@
 import { SafeString } from "../core/environment.ts";
 import type { Environment, Plugin } from "../core/environment.ts";
 
-const UNSAFE = /['"&<>]/g;
-const escapeMap: Record<string, string> = {
-  "'": "&apos;",
-  '"': "&quot;",
-  "&": "&amp;",
-  "<": "&lt;",
-  ">": "&gt;",
-};
+const UNSAFE = /[<>"&']/;
 
 export default function (): Plugin {
   return (env: Environment) => {
-    env.filters.escape = (value: unknown) => {
-      if (value instanceof SafeString) return value.toString();
+    env.filters.escape = (value: any): string => {
       if (!value) return "";
-      return value.toString().replace(UNSAFE, (match) => escapeMap[match]);
+      if (value instanceof SafeString) return value.toString();
+
+      const str = value.toString();
+
+      const match = UNSAFE.exec(str);
+
+      if (match === null) {
+        return str;
+      }
+
+      let html = '';
+
+      for (let i = match.index; i < str.length; i++) {
+        switch (str.charCodeAt(i)) {
+          case 34: // "
+            html += '&quot;';
+          break;
+          case 39: // '
+            html += '&apos;';
+          break;
+          case 38: // &
+            html += '&amp;';
+          break;
+          case 60: // <
+            html += '&lt;';
+          break;
+          case 62: // >
+            html += '&gt;';
+          break;
+          default:
+            html += str[i];
+        }
+      }
+
+      return html;
     };
   };
 }