This is for the v2 branch to avoid some conficts.

I noticed the benchmarks for Vento were with autoescape disabled, whereas most of the other template engines have it enabled by default. I found the performance of the escape to be pretty slow compared to PugJS (the only other fast templating engine) in particular, so this change fixes the performance.

Before:
```
benchmark         time/iter (avg)        iter/s      (min … max)           p75      p99     p995
----------------- ----------------------------- --------------------- --------------------------
Vento                    349.7 ns     2,860,000 (333.6 ns … 431.9 ns) 353.7 ns 414.4 ns 431.9 ns
VentoAutoescape            1.1 µs       895,800 (  1.1 µs …   1.3 µs)   1.1 µs   1.3 µs   1.3 µs
Pug                      677.4 ns     1,476,000 (665.0 ns … 746.1 ns) 676.5 ns 746.1 ns 746.1 ns
```

After:
```
benchmark         time/iter (avg)        iter/s      (min … max)           p75      p99     p995
----------------- ----------------------------- --------------------- --------------------------
Vento                    344.5 ns     2,902,000 (327.8 ns … 474.5 ns) 355.2 ns 408.5 ns 474.5 ns
VentoAutoescape          618.0 ns     1,618,000 (598.5 ns … 763.7 ns) 637.5 ns 763.7 ns 763.7 ns
Pug                      678.0 ns     1,475,000 (669.1 ns … 735.1 ns) 677.8 ns 735.1 ns 735.1 ns
```